// src/interfaces/Prices.ts
// Generated by https://quicktype.io

export interface Prices {
    default_credit: number;
    prices:         Price[];
}

export interface Price {
    rate:          number;
    imp_finanz:    number;
    imp_credito:   number;
    imp_rata:      number;
    inter:         number;
    price_credito: number;
    tan:           number;
    taeg:          number;
    erogato:       number;
    teg:           number;
    dovuto:        number;
    notes?:        string;
}